--- demo/src/main/java/com/example/demo/PatientManagementApplication.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PatientManagementApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PatientManagementApplication.class, args);
-	}
-
-}
