+++ demo/src/main/java/com/kajucode/patient/repository/entity/PatientEntity.java
@@ -1,6 +1,12 @@
-package com.example.demo.repository.entity;
+package com.kajucode.patient.repository.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
 import lombok.Getter;
 import lombok.Setter;
